+++ fhir_core/constraints.py
@@ -1,0 +1,5 @@
+__author__ = "Md Nazrul Islam"
+__email__ = "email2nazrul@gmail.com"
+
+TYPES_ID_MAX_LENGTH = 255
+TYPES_STRING_ALLOW_EMPTY_STR = False